Remove EncounterReference extension and update Consent resources IKNL/PZP-FHIR-R4#71
The custom EncounterReference extension and its references have been removed from Consent, Goal, and DeviceUseStatement resources. Consent resources have been updated to eliminate the extension and ensure correct placement of the policy field. The implementation guide index was updated to reflect the removal of the extension and to clarify the focus on the Communication profile and ACP Questionnaire.